refactor(ui): move the command-shell handoff out of model.go
model.go had grown past 2000 lines by accretion rather than structure. Split
it by theme, starting with the shell handoff, which shares no state with the
dashboard beyond suspending it.

Declarations move unchanged; this is a file-organisation change only.
